test: moved pda tests in dedicated class
